[BOOKINGSG-9154][IT] Rename style file extensions and update imports
